+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.1.0
+
+* serializers can now modify outgoing headers
+* HTTP methods (get, post, ...) never throw a HttpException,
+    not even with 500 http status response
+* method RestResult.assertSuccess() when you are not interested in response body (successData)        
+
 ## 3.0.0
 
 * More flexible serializers and deserializers. (_Breaking api change if you implement your own custom serializers or deserializers_)